Add navigation component and restructure About Us and Home pages with new sections
