+++ src/components/AboutUs/CTASection/index.tsx
@@ -1,0 +1,25 @@
+const CTASection = () => {
+  return (
+    <section className="py-20 bg-gray-50 w-full">
+      <div className="w-full px-4 sm:px-6 lg:px-8 text-center">
+        <h2 className="text-3xl md:text-4xl font-bold text-gray-900 mb-4">
+          Ready to Work With Us?
+        </h2>
+        <p className="text-xl text-gray-600 mb-8 max-w-3xl mx-auto">
+          Let's discuss how we can help you achieve your financial goals and
+          growth objectives.
+        </p>
+        <div className="flex flex-col sm:flex-row gap-4 justify-center">
+          <button className="bg-blue-900 text-white px-8 py-3 rounded-lg font-semibold hover:bg-blue-800 transition-colors">
+            Schedule a Consultation
+          </button>
+          <button className="bg-blue-900 text-white px-8 py-3 rounded-lg font-semibold hover:bg-blue-800 transition-colors">
+            Download Our Brochure
+          </button>
+        </div>
+      </div>
+    </section>
+  );
+};
+
+export default CTASection;
