+++ src/components/AboutUs/ValuesSection/index.tsx
@@ -1,0 +1,33 @@
+import { values } from "../../../utils/const";
+
+const ValuesSection = () => {
+  return (
+    <section className="py-20 bg-gray-50 w-full">
+      <div className="w-full px-4 sm:px-6 lg:px-8">
+        <div className="text-center mb-16">
+          <h2 className="text-3xl md:text-4xl font-bold text-gray-900 mb-4">
+            Our Values
+          </h2>
+          <p className="text-xl text-gray-600 max-w-3xl mx-auto">
+            The principles that guide everything we do
+          </p>
+        </div>
+        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-8">
+          {values.map((value, index) => (
+            <div key={index} className="text-center">
+              <div className="bg-blue-100 w-16 h-16 rounded-full flex items-center justify-center mx-auto mb-4">
+                <span className="text-2xl">{value.icon}</span>
+              </div>
+              <h3 className="text-xl font-semibold text-gray-900 mb-3">
+                {value.title}
+              </h3>
+              <p className="text-gray-600">{value.description}</p>
+            </div>
+          ))}
+        </div>
+      </div>
+    </section>
+  );
+};
+
+export default ValuesSection;
