+++ src/components/Home/HeroSection/index.tsx
@@ -1,0 +1,41 @@
+import { Link } from "react-router-dom";
+
+const HeroSection = () => {
+  return (
+    <section className="relative bg-gradient-to-br from-slate-900 via-blue-900 to-indigo-900 text-white w-full overflow-hidden">
+      {/* Background Pattern */}
+      <div className="absolute inset-0 bg-[url('data:image/svg+xml,%3Csvg%20width%3D%2260%22%20height%3D%2260%22%20viewBox%3D%220%200%2060%2060%22%20xmlns%3D%22http%3A//www.w3.org/2000/svg%22%3E%3Cg%20fill%3D%22none%22%20fill-rule%3D%22evenodd%22%3E%3Cg%20fill%3D%22%23ffffff%22%20fill-opacity%3D%220.05%22%3E%3Ccircle%20cx%3D%2230%22%20cy%3D%2230%22%20r%3D%222%22/%3E%3C/g%3E%3C/g%3E%3C/svg%3E')] opacity-20"></div>
+
+      {/* Floating Elements */}
+      <div className="absolute top-20 left-10 w-20 h-20 bg-blue-500/20 rounded-full blur-xl animate-pulse"></div>
+      <div className="absolute top-40 right-20 w-32 h-32 bg-indigo-500/20 rounded-full blur-2xl animate-pulse delay-1000"></div>
+      <div className="absolute bottom-20 left-1/4 w-24 h-24 bg-purple-500/20 rounded-full blur-xl animate-pulse delay-2000"></div>
+
+      <div className="relative w-full px-4 sm:px-6 lg:px-8 py-32">
+        <div className="text-center max-w-5xl mx-auto">
+          <div className="inline-flex items-center px-4 py-2 bg-white/10 backdrop-blur-sm rounded-full text-sm font-medium mb-8 border border-white/20">
+            <span className="w-2 h-2 bg-green-400 rounded-full mr-2 animate-pulse"></span>
+            Trusted by 500+ Companies Worldwide
+          </div>
+          <h1 className="text-5xl md:text-7xl font-black mb-8 bg-gradient-to-r from-white via-blue-100 to-indigo-200 bg-clip-text text-transparent leading-tight">
+            Expert Financial Advisory Services
+          </h1>
+          <p className="text-xl md:text-2xl mb-12 text-blue-100 max-w-3xl mx-auto leading-relaxed">
+            Strategic guidance for fundraising and financial advisory needs. We
+            help businesses navigate complex financial landscapes with
+            confidence.
+          </p>
+          <div className="flex flex-col sm:flex-row gap-6 justify-center items-center">
+            <Link to="/about">
+              <button className="bg-gradient-to-r from-blue-600 to-indigo-600 text-white group border-2 border-blue-500/50 px-10 py-4 rounded-xl font-semibold hover:from-blue-700 hover:to-indigo-700 hover:border-blue-400/70 transition-all duration-300 backdrop-blur-sm shadow-lg hover:shadow-xl">
+                <span className="flex items-center">Learn More</span>
+              </button>
+            </Link>
+          </div>
+        </div>
+      </div>
+    </section>
+  );
+};
+
+export default HeroSection;
