+++ src/components/Home/WhyChooseUsSection/index.tsx
@@ -1,0 +1,51 @@
+const WhyChooseUsSection = () => {
+  return (
+    <section className="py-20 bg-gray-50 w-full">
+      <div className="w-full px-4 sm:px-6 lg:px-8">
+        <div className="text-center mb-16">
+          <h2 className="text-3xl md:text-4xl font-bold text-gray-900 mb-4">
+            Why Choose Upama Financial?
+          </h2>
+        </div>
+        <div className="grid grid-cols-1 md:grid-cols-3 gap-8">
+          <div className="text-center">
+            <div className="bg-blue-100 w-16 h-16 rounded-full flex items-center justify-center mx-auto mb-4">
+              <span className="text-2xl">🎯</span>
+            </div>
+            <h3 className="text-xl font-semibold text-gray-900 mb-2">
+              Expertise
+            </h3>
+            <p className="text-gray-600">
+              Deep industry knowledge and proven track record in financial
+              advisory
+            </p>
+          </div>
+          <div className="text-center">
+            <div className="bg-blue-100 w-16 h-16 rounded-full flex items-center justify-center mx-auto mb-4">
+              <span className="text-2xl">⚡</span>
+            </div>
+            <h3 className="text-xl font-semibold text-gray-900 mb-2">
+              Efficiency
+            </h3>
+            <p className="text-gray-600">
+              Streamlined processes and quick turnaround times for all services
+            </p>
+          </div>
+          <div className="text-center">
+            <div className="bg-blue-100 w-16 h-16 rounded-full flex items-center justify-center mx-auto mb-4">
+              <span className="text-2xl">🤝</span>
+            </div>
+            <h3 className="text-xl font-semibold text-gray-900 mb-2">
+              Partnership
+            </h3>
+            <p className="text-gray-600">
+              Long-term relationships built on trust and mutual success
+            </p>
+          </div>
+        </div>
+      </div>
+    </section>
+  );
+};
+
+export default WhyChooseUsSection;
