+++ src/utils/const.ts
@@ -1,0 +1,78 @@
+export const fundraisingServices = [
+  {
+    title: "Venture Capital (VC)",
+    description:
+      "Strategic investment in early-stage companies with high growth potential",
+    icon: "🚀",
+  },
+  {
+    title: "Private Equity (PE)",
+    description:
+      "Investment in mature companies to drive operational improvements and growth",
+    icon: "💼",
+  },
+  {
+    title: "Pre-IPO Advisory",
+    description: "Preparation and guidance for companies planning to go public",
+    icon: "📈",
+  },
+  {
+    title: "IPO Services",
+    description:
+      "Complete support for Initial Public Offering process and execution",
+    icon: "🏛️",
+  },
+  // {
+  //   title: "Sell Side Advisory",
+  //   description:
+  //     "Strategic guidance for companies looking to sell or divest assets",
+  //   icon: "💰",
+  // },
+];
+
+export const advisoryServices = [
+  // {
+  //   title: "M&A Advisory",
+  //   description: "Comprehensive merger and acquisition advisory services",
+  //   icon: "🤝",
+  // },
+  {
+    title: "Sell Side Advisory",
+    description:
+      "Strategic guidance for companies looking to sell or divest assets",
+    icon: "💰",
+  },
+  {
+    title: "Buy Side Advisory",
+    description:
+      "Strategic guidance for companies seeking acquisition opportunities",
+    icon: "🎯",
+  },
+];
+
+export const values = [
+  {
+    title: "Integrity",
+    description:
+      "We maintain the highest ethical standards in all our advisory relationships",
+    icon: "🏛️",
+  },
+  {
+    title: "Excellence",
+    description:
+      "We deliver superior results through deep expertise and attention to detail",
+    icon: "⭐",
+  },
+  {
+    title: "Innovation",
+    description:
+      "We leverage cutting-edge approaches to solve complex financial challenges",
+    icon: "💡",
+  },
+  {
+    title: "Partnership",
+    description:
+      "We build long-term relationships based on trust and mutual success",
+    icon: "🤝",
+  },
+];
