+++ tsconfig.app.json
@@ -23,5 +23,5 @@
     "noFallthroughCasesInSwitch": true,
     "noUncheckedSideEffectImports": true
   },
-  "include": ["src", "src/components/Navigation.tsx"]
+  "include": ["src", "index.tsx", "src/components/Home/HeroSection"]
 }
